Refactor linumpy module structure and update imports
- Moved `data_io` from `linumpy.utils` to `linumpy.io`.
- Introduced `mosaic_grid.py` in `linumpy.stitching` for managing mosaic grid images.
- Updated imports in various scripts and test files to reflect the new module structure.
- Removed unused imports from `linumpy.utils.__init__.py`.
- Adjusted references to `MosaicGrid` and related functions in scripts for illumination compensation, cropping, and stitching.
- Ensured consistency in the import paths across the codebase.
