+++ linumpy/io/__init__.py
@@ -1,3 +1,4 @@
 from .allen import *
+from .data_io import *
 from .zarr import *
 from .npz import write_numpy, read_numpy, read_numpy_data, read_numpy_metadata
