+++ linumpy/shifts/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .utils import *  # noqa: F401, F403
