+++ linumpy/shifts/utils.py
