+++ linumpy/stitching/__init__.py
@@ -1,3 +1,4 @@
 from .FileUtils import *
+from .mosaic_grid import *
 from .stitch_utils import *
 from .topology import *
