+++ linumpy/stitching/mosaic_grid.py
