--- linumpy/utils/__init__.py
@@ -1,2 +1,0 @@
-from .data_io import *
-from .mosaic_grid import *
