+++ scripts/linum_crop_tiles.py
@@ -7,7 +7,7 @@ import argparse
 from pathlib import Path
 import SimpleITK as sitk
 
-from linumpy.utils.mosaic_grid import MosaicGrid
+from linumpy.stitching.mosaic_grid import MosaicGrid
 
 
 def _build_arg_parser():
