+++ scripts/linum_stitch_2d.py
@@ -10,7 +10,7 @@ from pathlib import Path
 import SimpleITK as sitk
 import numpy as np
 
-from linumpy.utils.mosaic_grid import MosaicGrid
+from linumpy.stitching.mosaic_grid import MosaicGrid
 
 
 def _build_arg_parser():
